requirements.txt
langchain
openai
python-dotenv
langchain-openai
langchain-core
langchain-community
tavily-python
langgraph
langgraph-prebuilt
langgraph-sdk
langgraph-checkpoint-sqlite
ipython
langgraph-cli
langgraph-cli[inmem]
langchain-tavily
wikipedia
trustcall